requirements.txt
asgiref==3.4.1
auth0-python==3.19.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.0
django-cors-headers==3.11.0
djangorestframework==3.13.1
idna==3.3
numpy==1.22.1
oauthlib==3.1.1
Pillow==9.0.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.21
PyJWT==2.3.0
python-decouple==3.5
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-core==4.1.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
